Move model.Location to data package
